# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GrantLayer MVP API
  description: |
    GrantLayer MVP — API Contract Baseline (GL-203B).

    **Status: Developer Preview / Controlled Preview with strict boundaries.**
    This is not a production SaaS. Not ready for real customer data, private
    grant data, or institutional data. Tenant/workspace isolation baseline is
    implemented but not production-complete. No official SDK/package is claimed
    or published.

    This document describes the HTTP surface as of GL-203B, aligned with the
    actual implementation after GL-200 through GL-203.

    **Auth modes:**
    - Legacy mode (`ENABLE_OPERATOR_MODEL=false`): `Authorization: Bearer <admin-token>`
      via constant-time HMAC comparison. Tenant context is bound to `"demo"`.
    - Operator mode (`ENABLE_OPERATOR_MODEL=true`): `Authorization: Bearer <operator-token>`
      via PBKDF2-HMAC-SHA256 (600,000 iterations). Tenant context is derived
      server-side from the authenticated operator record.

    **Tenant context:**
    `tenant_id` is always derived server-side from authentication. Clients cannot
    supply or override `tenant_id` via headers, path parameters, or request bodies.
    Cross-tenant ID lookups return HTTP 404; cross-tenant list endpoints return
    an empty list. No existence of cross-tenant resources is leaked.

    **workspace_id:**
    `workspace_id` is reserved, nullable, and not currently enforced. It is present
...

    **Correlation ID:**
...

...

...

...
    ```json
    {
...

...

...
servers:
...
paths:
...
  /v1/exercise:
    post:
      summary: Exercise a grant (record an authorization decision)
      description: |
        Evaluates the request against the caller's workspace grants and
        records the decision: a GrantExecution row and a chain-linked
        AuditEvent. The endpoint performs no action itself — the agent acts
        on its own side once (and only if) the decision is "allowed".
        The former path /v1/demo-action keeps answering as a 307 alias.
        Authorization: owner, grant_admin (operator mode) or admin token (legacy mode).
      operationId: exercise
      security:
        - LegacyAdminToken: []
        - OperatorToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subjectId
                - role
                - action
                - resource
              properties:
                subjectId: { type: string }
                role: { type: string }
                action: { type: string }
                resource: { type: string }
                challengeId: { type: string, nullable: true }
                attemptedFeeLovelace:
                  type: integer
                  nullable: true
                  minimum: 0
                  description: |
                    Declared fee attempt in lovelace for fee-constrained
                    grants. Optional against unconstrained grants; a
                    fee-constrained grant denies fail-closed
                    (constraint_attempt_undeclared) when omitted, and denies
                    constraint_violated_max_fee when the declaration exceeds
                    the signed limit (the violation is witnessed on the audit
                    chain as {"type","limit","attempted"}).
      responses:
        "200":
          description: Exercise allowed (decision recorded)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseResult"
        "400":
          description: Invalid JSON or missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authorization failed or action denied by policy engine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/demo/tamper-grant/{id}:
    post:
      summary: Demo tamper a grant (DEMO ONLY)
      description: |
        **Demo endpoint.** Corrupts a grant field without re-signing to
        demonstrate tamper detection. Requires `ENABLE_DEMO_ENDPOINTS=true`.
        Never available in production builds.
      operationId: tamperGrant
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      security:
        - LegacyAdminToken: []
        - OperatorToken: []
      responses:
        "200":
          description: Tamper result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  grant_id: { type: string }
                  field_changed: { type: string }
        "403":
          description: Demo endpoints disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Grant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    LegacyAdminToken:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Legacy admin token mode. Send `Authorization: Bearer <admin-token>`.
        Used when `ENABLE_OPERATOR_MODEL=false`. Validated via constant-time
        HMAC comparison (GL-201). Tenant context is bound to `"demo"` server-side.
        Clients cannot supply or override tenant_id. In production-like runtime
        mode, placeholder/weak tokens are rejected at startup.
    OperatorToken:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Operator bearer token mode. Send `Authorization: Bearer <operator-token>`.
        Used when `ENABLE_OPERATOR_MODEL=true`. Validated via PBKDF2-HMAC-SHA256
        (600,000 iterations) with SHA-256 lookup-hash pre-filter. Tenant context
        is derived server-side from the authenticated operator record. Clients
        cannot supply or override tenant_id. Tokens are issued once at operator
        creation and are not retrievable afterward.
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error summary
        errorCode:
          type: string
          description: Machine-readable error code (GL-030 additive field)
        reason:
          type: string
          description: Detailed explanation (GL-030 additive field)
      example:
        error: "Grant execution not found"
        errorCode: "grant_execution_not_found"
        reason: "The requested grant execution does not exist."
...
    ExerciseResult:
      type: object
      properties:
        approved: { type: boolean }
        result: { type: string, enum: [allowed, denied, failed] }
        reasonCode: { type: string, nullable: true }
        message: { type: string, nullable: true }
        reason: { type: string, nullable: true }
        matchedGrantId: { type: string, nullable: true }
        challengeId: { type: string, nullable: true }
        challengeResult: { type: string, nullable: true }
        auditEventId: { type: string }
        grantSignatureResult: { type: string }
        executionId: { type: string }
...